+++ js/app/packages/app/component/Layout.tsx
@@ -8,11 +8,9 @@ import {
   LAYOUT_CONTEXT_ID,
   setPersistedLayoutSizes,
 } from '@core/signal/layout';
-import { isEditableInput } from '@core/util/isEditableInput';
-import { isIOS } from '@solid-primitives/platform';
 import { type RouteSectionProps, useLocation } from '@solidjs/router';
 import { attachGlobalDOMScope } from 'core/hotkey/hotkeys';
-import { createEffect, onCleanup, onMount, Show, Suspense } from 'solid-js';
+import { createEffect, onMount, Show, Suspense } from 'solid-js';
 import { updateCookie } from '../util/updateCookie';
 import Banner from './banner/Banner';
 import { GlobalBulkEditEntityModal } from './bulk-edit-entity/BulkEditEntityModal';
@@ -25,12 +23,9 @@ import { QuickCreateMenu } from './QuickCreateMenu';
 import { RightbarWrapper } from './rightbar/Rightbar';
 import { SettingsWrapper } from './settings/SettingsWrapper';
 import { ShortcutsHelper } from './settings/ShortcutsHelper';
-import { isNativeMobilePlatform } from '@core/mobile/isNativeMobilePlatform';
-import {
-  setVirtualKeyboardVisible,
-  virtualKeyboardVisible,
-} from '@core/mobile/virtualKeyboard';
+import { virtualKeyboardVisible } from '@core/mobile/virtualKeyboard';
 import { cn } from '@ui/utils/classname';
+import { useAppSquishHandlers } from './useAppSquishHandlers';
 
 const AUTH_URLS = [
   '/app/login',
@@ -45,6 +40,8 @@ export function Layout(props: RouteSectionProps) {
   const isAuthenticated = useIsAuthenticated();
   const { paywallOpen, showPaywall } = usePaywallState();
   const location = useLocation();
+
+  useAppSquishHandlers();
 
   // save last_path to cookie
   createEffect(() => {
@@ -61,88 +58,6 @@ export function Layout(props: RouteSectionProps) {
       sameSite: 'Lax',
     });
   });
-
-  if (isIOS) {
-    // We are tracking viewport height, and using that to set a CSS variable and the viewport offset, so that we can properly constrain the viewport-height for mobile in response to changes such as the virtual keyboard appearing
-    let previousViewportHeight = window.visualViewport?.height || 0;
-    const handleResize = () => {
-      if (window.visualViewport) {
-        const newViewportHeight = window.visualViewport.height;
-        if (newViewportHeight < previousViewportHeight) {
-          setVirtualKeyboardVisible(true);
-          const vh = newViewportHeight * 0.01;
-          document.documentElement.style.setProperty('--dvh', `${vh}px`);
-          setTimeout(() => {
-            window.scrollTo(0, 0);
-          });
-        } else {
-          setVirtualKeyboardVisible(false);
-          document.documentElement.style.setProperty('--dvh', '1dvh');
-        }
-        previousViewportHeight = newViewportHeight;
-      }
-    };
-
-    const handleFocusOut = (e: FocusEvent) => {
-      if (
-        e.target instanceof Element &&
-        isEditableInput(e.target) &&
-        (!e.relatedTarget ||
-          (e.relatedTarget instanceof Element &&
-            !isEditableInput(e.relatedTarget)))
-      ) {
-        document.documentElement.style.setProperty('--dvh', '1dvh');
-        setVirtualKeyboardVisible(false);
-      }
-    };
-
-    onMount(() => {
-      if (window.visualViewport) {
-        window.visualViewport.addEventListener('resize', handleResize);
-        handleResize();
-        window.visualViewport.addEventListener('scroll', handleResize);
-      }
-      document.addEventListener('focusout', handleFocusOut);
-
-      onCleanup(() => {
-        if (window.visualViewport) {
-          window.visualViewport.removeEventListener('resize', handleResize);
-          window.visualViewport.removeEventListener('scroll', handleResize);
-        }
-        document.removeEventListener('focusout', handleFocusOut);
-      });
-    });
-  }
-
-  if (isNativeMobilePlatform()) {
-    type VirtualKeyboardEvent = CustomEventInit<{
-      height: number;
-      duration: number;
-    }>;
-
-    const handleKeyboardWillShow = (event: VirtualKeyboardEvent) => {
-      setVirtualKeyboardVisible(true);
-      const newViewportHeight =
-        (window.visualViewport?.height ?? 0) - (event.detail?.height ?? 0);
-      const vh = newViewportHeight * 0.01;
-      document.documentElement.style.setProperty('--dvh', `${vh}px`);
-    };
-
-    const handleKeyboardWillHide = () => {
-      setVirtualKeyboardVisible(false);
-      document.documentElement.style.setProperty('--dvh', '1dvh');
-    };
-
-    onMount(() => {
-      window.addEventListener('keyboardWillShow', handleKeyboardWillShow);
-      window.addEventListener('keyboardWillHide', handleKeyboardWillHide);
-
-      onCleanup(() => {
-        window.removeEventListener('keyboardWillShow', handleKeyboardWillShow);
-        window.removeEventListener('keyboardWillHide', handleKeyboardWillHide);
-      });
-    });
-  }
 
   onMount(() => {
     if (sessionStorage.getItem('showUpgradeModal') === 'true') {
